Updated service types for compatibility with 0.9.0
Has some debugging code.

Right now the print_joint_poses example freezes.

There are various status enums in version 0.9.0 that have some statuses
that have been removed from this later version.

Meanwhile there are extra statuses added into the middle of the enums.

I suspect that there is a different status numbering scheme in the protocols and therefore I tried to make the enums work such that

1. All statuses shared between 0.9.0 and later versions have the same value as they had in 0.9.0.
2. All statuses that were removed from 0.9.0 are present
3. Statuses added in the later version are still present (so everything compiles), but the value is now greater than all 0.9.0 status.
There is an assumption that we won't ever see the new status with the old robot.